+++ FoodBowl/Global/View/Component/CustomButton/SearchBarButton.swift
@@ -13,7 +13,7 @@ import Then
 final class SearchBarButton: UIButton {
     // MARK: - property
     private let searchIconView = UIImageView().then {
-        $0.image = ImageLiteral.search.resize(to: CGSize(width: 20, height: 20))
+        $0.image = ImageLiteral.search.resize(to: CGSize(width: 20, height: 20)).withRenderingMode(.alwaysTemplate)
         $0.tintColor = .subText
     }
