+++ FoodBowl/Screens/Common/NewFeed/View/SetStoreViewController.swift
@@ -22,78 +22,94 @@ final class SetStoreViewController: BaseViewController {
         fatalError("init(coder:) has not been implemented")
     }
 
+    private var stores = [Place]()
+
     // MARK: - property
-    private let guideLabel = UILabel().then {
-        $0.text = "후기를 작성할 가게를 찾아주세요."
-        $0.font = UIFont.preferredFont(forTextStyle: .body, weight: .medium)
-        $0.textColor = .mainText
+    private lazy var searchController = UISearchController(searchResultsController: nil).then {
+        $0.searchBar.placeholder = "검색"
+        $0.searchResultsUpdater = self
+        $0.obscuresBackgroundDuringPresentation = false
     }
 
-    private lazy var searchBarButton = SearchBarButton().then {
-        $0.placeholderLabel.text = "가게 검색"
-        let action = UIAction { [weak self] _ in
-            guard let viewModel = self?.viewModel else { return }
-            let searchStoreViewController = SearchStoreViewController(viewModel: viewModel)
-            let navigationController = UINavigationController(rootViewController: searchStoreViewController)
-            navigationController.modalPresentationStyle = .fullScreen
-            searchStoreViewController.delegate = self
-            DispatchQueue.main.async {
-                self?.present(navigationController, animated: true)
-            }
+    private lazy var cancelButton = UIButton().then {
+        $0.setTitle("취소", for: .normal)
+        $0.setTitleColor(.mainPink, for: .normal)
+        $0.titleLabel?.font = UIFont.preferredFont(forTextStyle: .headline, weight: .regular)
+        let buttonAction = UIAction { [weak self] _ in
+            self?.dismiss(animated: true, completion: nil)
         }
-        $0.addAction(action, for: .touchUpInside)
+        $0.addAction(buttonAction, for: .touchUpInside)
     }
 
-    private lazy var selectedStoreView = SelectedStoreView().then {
-        $0.isHidden = true
+    private lazy var storeInfoTableView = UITableView().then {
+        $0.register(StoreSearchTableViewCell.self, forCellReuseIdentifier: StoreSearchTableViewCell.className)
+        $0.delegate = self
+        $0.dataSource = self
+        $0.separatorInset = UIEdgeInsets(top: 0, left: 0, bottom: 20, right: 0)
+        $0.backgroundColor = .clear
     }
 
     // MARK: - life cycle
+
     override func setupLayout() {
-        view.addSubviews(guideLabel, searchBarButton, selectedStoreView)
-
-        guideLabel.snp.makeConstraints {
-            $0.top.equalTo(view.safeAreaLayoutGuide.snp.top).inset(20)
-            $0.leading.equalToSuperview().inset(BaseSize.horizantalPadding)
+        view.addSubviews(storeInfoTableView)
+        storeInfoTableView.snp.makeConstraints {
+            $0.top.equalTo(view.safeAreaLayoutGuide)
+            $0.leading.trailing.bottom.equalToSuperview()
         }
+    }
 
-        searchBarButton.snp.makeConstraints {
-            $0.top.equalTo(guideLabel.snp.bottom).offset(20)
-            $0.leading.trailing.equalToSuperview().inset(BaseSize.horizantalPadding)
-            $0.height.equalTo(40)
+    override func setupNavigationBar() {
+        super.setupNavigationBar()
+        navigationItem.searchController = searchController
+        title = "맛집 검색"
+    }
+
+    private func searchStores(keyword: String) {
+        Task {
+            stores = await viewModel.searchStores(keyword: keyword)
+            storeInfoTableView.reloadData()
         }
+    }
 
-        selectedStoreView.snp.makeConstraints {
-            $0.top.equalTo(searchBarButton.snp.bottom).offset(10)
-            $0.leading.trailing.equalToSuperview().inset(BaseSize.horizantalPadding)
-            $0.height.equalTo(60)
+    private func setStore(store: Place) {
+        Task {
+            await viewModel.setStore(store: store)
+            navigationController?.popViewController(animated: true)
         }
     }
 }
 
-extension SetStoreViewController: SearchStoreViewControllerDelegate {
-    func setStore(store: Place) {
-        searchBarButton.placeholderLabel.text = "가게 재검색"
-        selectedStoreView.storeNameLabel.text = store.placeName
-        selectedStoreView.storeAdressLabel.text = store.addressName
-        selectedStoreView.isHidden = false
+extension SetStoreViewController: UITableViewDataSource, UITableViewDelegate {
+    func tableView(_: UITableView, numberOfRowsInSection _: Int) -> Int {
+        return stores.count
+    }
 
-        let action = UIAction { [weak self] _ in
-            let showWebViewController = ShowWebViewController()
-            showWebViewController.title = "가게 정보"
-            showWebViewController.url = store.placeURL
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard let cell = tableView
+            .dequeueReusableCell(withIdentifier: StoreSearchTableViewCell.className, for: indexPath) as? StoreSearchTableViewCell
+        else { return UITableViewCell() }
 
-            let navigationController = UINavigationController(rootViewController: showWebViewController)
-            navigationController.modalPresentationStyle = .fullScreen
+        cell.selectionStyle = .none
+        cell.storeNameLabel.text = stores[indexPath.item].placeName
+        cell.storeAdressLabel.text = stores[indexPath.item].addressName
+        cell.storeDistanceLabel.text = stores[indexPath.item].distance.prettyDistance
 
-            DispatchQueue.main.async {
-                self?.present(navigationController, animated: true)
-            }
-        }
-        selectedStoreView.mapButton.addAction(action, for: .touchUpInside)
+        return cell
+    }
 
-        Task {
-            await viewModel.setStore(store: store)
-        }
+    func tableView(_: UITableView, heightForRowAt _: IndexPath) -> CGFloat {
+        return 60
+    }
+
+    func tableView(_: UITableView, didSelectRowAt indexPath: IndexPath) {
+        setStore(store: stores[indexPath.item])
+    }
+}
+
+extension SetStoreViewController: UISearchResultsUpdating {
+    func updateSearchResults(for searchController: UISearchController) {
+        guard let text = searchController.searchBar.text?.lowercased() else { return }
+        searchStores(keyword: text)
     }
 }
